+++ sync-for-payroll/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] sync-for-payroll
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-sync-for-payroll/0.1.0 - sync-for-payroll
+
+## 2023-09-08 00:09:55
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.82.0 (2.107.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.0] sync-for-payroll
+### Releases
+- [PyPI v0.2.0] https://pypi.org/project/codat-sync-for-payroll/0.2.0 - sync-for-payroll